refactor: move binding.py into builder/_binding.py
binding + pointer utilities are internal to the builder, not
user-facing API. Move inside the builder package.
