+++ tests/test_binding.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 
 from genro_bag import Bag
 
-from genro_builders.binding import BindingManager
+from genro_builders.builder._binding import BindingManager
 from genro_builders.builder_bag import BuilderBag
