+++ tests/test_compiler_base.py
@@ -10,7 +10,7 @@ from genro_bag import Bag
 
 from genro_builders import BagBuilderBase, BagCompilerBase
 from genro_builders.compiler import compiler
-from genro_builders.binding import BindingManager
+from genro_builders.builder._binding import BindingManager
 from genro_builders.builder_bag import BuilderBag
 from genro_builders.builder import component, element
